Refactor test-only functionality
